requirements.txt
aiosmtplib==3.0.2
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.2.1
black==24.10.0
blinker==1.9.0
bump-pydantic==0.8.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-mail==1.4.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.5
libcst==1.6.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pyasn1==0.6.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typer==0.15.1
typing_extensions==4.12.2
uvicorn==0.34.0
